Added HTTP logging for error check
Added UI for showing weather and forecast list
Had to change the forecast API -> @GET("forecast") to @GET("forecast/daily") to get daily API
Updated Models for forecase API
Added extensions for date conversion
Added Necessary Fonts
Added Landscape view for when in landscape

+++ app/src/main/java/com/textsdev/cleanweatherapp_mvvm/data/model/WeatherForecastOfflineModel.kt
@@ -2,5 +2,5 @@ package com.textsdev.cleanweatherapp_mvvm.data.model
 
 data class WeatherForecastOfflineModel(
     val date: Int,
-    val temp: Double
+    val temp: Int
 )
